Added location search
The map with a search bar turned out to be FAR more difficult to implement on iOS than in Kotlin when using the Swift interface instead of Storyboard.  Since being able to click on a location on a map is not as helpful as the search feature anyway, I just added the search feature on its own. There are currently several artifacts from the process, which were kept because we may include them at a later date