Bugfixes
Fixed accidentally uncommenting experimental persistent data mechanisms that did not work.
Fixed time getting reset to the current time when editing an event that already had a start time.
Fixed eventList changes when logging in. It now properly uploads any events made by the user before logging in to the database then downloads the database events. It is also all moved to one location instead of needing to be done in multiple locations