Citations & Bugfixes & Polish
Links are provided to the source of certain sections of code that I forgot to provide links to earlier. None of it is outright copied, but some of it is fairly similar. A few bugs were fixed (For example, drag & drop in the list view), and some possible polish was added